Revised message and market state data structures for better grouping and clarity. Updated GUI components to align with new structures and added header styling for enhanced visuals. Adjusted event processing logic to filter specific channels and ensured string inputs are validated properly.

+++ src/event/personal/status.rs
@@ -10,7 +10,7 @@ pub struct Status {
     pub flags: u32,
 
     #[serde(rename = "Flags2")]
-    pub flags2: Option<u16>,
+    pub flags2: Option<u32>,
 
     #[serde(rename = "Oxygen")]
     pub oxygen: Option<f64>,
